rewrite pet cmd to ts
